extract ValidationError into its own file
